3.4-django-testing/django_testing/requirements.txt
django~=4.2.3
djangorestframework~=3.14.0
django-filter~=23.2
psycopg2-binary
python-dotenv